Fix bug when opening URLs that StopTheNews can't handle.
Since version 4.0 of StopTheNews, we spoof the bundle identifier "com.apple.news",
so opening a URL with that bundle identifier doesn't open it in the News app,
it opens it in StopTheNews again! Instead we have to use the file system path
of the News app, which is /System/Applications/News.app in the Big Sur system volume.